+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `run_cpp_tests()` no longer accidentally reports that a test has been skipped (#2315).
 * `expect_setequal()` uses better wording in the results (@mcol, #2310).
+* `expect_shape()` evaluates `object` only once (@michaelchirico, #2345).
 
 # testthat 3.3.2